refactor: rename chat-panel.js to ChatPanel.js and fix import order
- Rename src/components/chat-panel.js to ChatPanel.js (PascalCase convention)
- Move createPortal import from react-dom to after @wordpress/* imports,
  under a separate 'External dependencies' comment block
- Update all import references from '../components/chat-panel' to
  '../components/ChatPanel' in admin-page, floating-widget, and screen-meta
- Update test file import and comment reference

Fixes #875
