+++ src/screen-meta/index.js
@@ -21,7 +21,7 @@ import STORE_NAME from '../store';
 // Register gratis-ai-agent-js/* client-side abilities into core/abilities
 // before the chat mounts (t165 — closes the wiring gap in #815).
 import '../abilities';
-import ChatPanel from '../components/chat-panel';
+import ChatPanel from '../components/ChatPanel';
 import './style.css';
 
 /**
